test: extract globals & use in test-abort-backtrace
First steps towards not needing to require test/common/index.js
everywhere.

Refs: https://github.com/nodejs/node/issues/20128#issuecomment-389933546
