+++ test/sequential/test-module-loading.js
@@ -252,6 +252,7 @@ try {
   assert.deepStrictEqual(children, {
     'common/index.js': {
       'common/fixtures.js': {},
+      'common/globals.js': {},
       'common/tmpdir.js': {}
     },
     'fixtures/not-main-module.js': {},
